better music search implementation

+++ Commands/download.js
@@ -1,21 +1,23 @@
-const {getYTPlaylist,getYTPlaylistID,getYTVideoID,getYTVideo,getVideoSearch,downloadAudio} = require("../YoutubeFunctions")
+const {getVideoSearch,downloadAudio} = require("../YoutubeFunctions")
 const fs = require("fs")
 module.exports.run =async (client, channel, authorID, args) => {
     return new Promise(function (resolve,reject){
-        getSong(args).then((Music)=>{
-            console.log(Music)
-            downloadAudio(Music.MusicUrl).then((FileName)=>{
-                console.log(Music.MusicTitle+" was downloaded")
-                channel.send(Music.MusicTitle+" was downloaded",{
-                    files:[
-                        "./Downloads/"+FileName
-                    ]
-                }).then(()=>{
-                    fs.unlinkSync("./Downloads/"+FileName)
+        getVideoSearch(args).then((Result)=>{
+            if (Result.Musics.length == 1) {
+                resolve("Trying to download your song please wait...")
+                downloadAudio(Result.Musics[0].MusicUrl).then((FileName)=>{
+                    channel.send(Result.Musics[0].MusicTitle+" was downloaded",{
+                        files:[
+                            "./Downloads/"+FileName
+                        ]
+                    }).then(()=>{
+                        fs.unlinkSync("./Downloads/"+FileName)
+                    })
                 })
-            })
+            } else {
+                resolve("Sorry, the bot does not support playlist downloading for now");
+            }
         })
-        resolve("Trying to download your song please wait...")
     })
 };
 module.exports.help = {
@@ -28,20 +30,3 @@ module.exports.help = {
         "type":3
     }]
 };
-async function getSong(args){
-    return new Promise(function (resolve, reject){
-        if(args[0]==undefined){
-            resolve("You need to specify a music to download")
-        } else if(args[0].includes("list=")){
-            resolve("The bot doesn't handle playlist downloading for now")
-        } else if(args[0].includes("watch?v=")){
-            getYTVideo(getYTVideoID(args[0])).then((res)=>{
-                resolve(res)
-            })
-        } else {
-            getVideoSearch(args).then((res)=>{
-                resolve(res)
-            })
-        }
-    })
-}
